refactor: 낙관적업데이트 적용오류 해결
